Fix linting crashing if a file to lint doesn't exist
This can occurr if for some reson there is no file at the give path.

In such a case SwiftLint will write to stderr:

```
Could not read contents of `path/to/file`
No lintable files found at path 'path/to/file'
```

and exit with code 1.

To our code this would simply look like an empty result, which would
then become an empty string, which cannot be parsed into a JSON object.