Don't lint files from excluded list in config YAML
Because our plugin runs on all the files added or modified by the PR,
unless a directory is given, we end up ignoring the "excluded"
configuration that users might have in the `.swiftlint.yml`.

This can be problematic in projects that track things like the
Carthage/Checkout folder in the repo. Two reasons:

- First of all the repos is set to exclude such directories, we should do the same; also
- A PR that updates one or more of these might end up with a lot of SwiftLint warnings. This in turn might result in Danger comment making a **long** comment, exceeding the 65536 characters limit imposed by the GitHub APIs.